+++ ServerCoreDefinitions/Operations/FileOperations.cs
@@ -162,7 +162,7 @@ namespace EasyITCenter.ServerCoreStructure {
         /// <param name="content"></param>
         /// <param name="rewrite"></param>
         public static void WriteToFile(string filePath, string content, bool rewrite = true) {
-            string[] pathParts = FileOperations.GetDirectoryFromFilePath(file).Split(Path.DirectorySeparatorChar);
+            string[] pathParts = FileOperations.GetDirectoryFromFilePath(filePath).Split(Path.DirectorySeparatorChar);
             for (int i = 0; i < pathParts.Length; i++) {
                 if (i > 0) { pathParts[i] = Path.Combine(pathParts[i - 1], pathParts[i]); }
                 if (!Directory.Exists(pathParts[i])) { Directory.CreateDirectory(pathParts[i]); }